+++ src/super/i-input/CHANGELOG.md
@@ -26,7 +26,7 @@ Changelog
 * Implemented new API from `iAccess`
 * Now `formValueConverter` can be provided as an array
 * Added support for the `Maybe` structure
-* Added the `attrs` prop
+* Added `attrsProp/attrs` properties
 * Added the `normalizeAttrs` method
 * Added the `nativeInput` block